Update test.py
This code is testing a smart contract called Trip. The Trip contract has several functions that can be called to interact with it.

The first part of the test code deploys the contract using a specific Ethereum address (0x5f4eC3Df9cbd43714FE2740f5E3616155c5b8419) as a parameter and the first account in the local Ganache instance (accounts[0]) as the deployer.

The following assertions check the initial state of the contract to ensure it was properly deployed. These assertions check values such as basePrice, lastPrice, nextRebaseTime, decimals, and maxSupply.

Next, the getPrice() function is tested to ensure that it returns the expected result. The updateBasePrice() function is then called and tested to ensure that the basePrice of the contract has been updated.

The calculateRebaseInterval() function is then tested to ensure that it returns one of the expected values.

Finally, the rebase() function is tested. This function is expected to fail and raise an error message because the maximum supply of the contract has been reached. The totalSupply() function is then called to ensure that the total supply of the contract hasn't changed. The nextRebaseTime() function is also checked to ensure it has been properly updated after the rebase attempt.